fix(ssr): fix expression scoping and other small fixes
