+++ packages/@lwc/engine-server/src/__tests__/fixtures/very-deeply-nested-each/expected.html
@@ -1,0 +1,7 @@
+<x-component>
+  <template shadowrootmode="open">
+    <div data-value="yolo">
+      yolo
+    </div>
+  </template>
+</x-component>
