+++ packages/@lwc/engine-server/src/__tests__/fixtures/very-deeply-nested-each/index.js
@@ -1,0 +1,3 @@
+export const tagName = 'x-component';
+export { default } from 'x/component';
+export * from 'x/component';
