+++ packages/@lwc/engine-server/src/__tests__/fixtures/very-deeply-nested-each/modules/x/component/component.html
@@ -1,0 +1,5 @@
+<template>
+	<template for:each={rows} for:item="row">
+		<div key={row.id} data-value={row.foo.bar.baz}>{row.foo.bar.baz}</div>
+	</template>
+</template>
