+++ packages/@lwc/engine-server/src/__tests__/fixtures/very-deeply-nested-each/modules/x/component/component.js
@@ -1,0 +1,14 @@
+import { LightningElement } from 'lwc';
+
+export default class Component extends LightningElement {
+  rows = [
+    {
+      id: 1,
+      foo: {
+        bar: {
+          baz: 'yolo'
+        }
+      }
+    }
+  ]
+}
