+++ packages/@lwc/ssr-compiler/src/compile-template/transformers/for-each.ts
@@ -8,25 +8,11 @@
 import { builders as b, is } from 'estree-toolkit';
 import { esTemplate } from '../../estemplate';
 import { irToEs } from '../ir-to-es';
-import { optimizeAdjacentYieldStmts } from '../shared';
+import { getScopedExpression, optimizeAdjacentYieldStmts } from '../shared';
 
 import type { ForEach as IrForEach } from '@lwc/template-compiler';
-import type {
-    Expression as EsExpression,
-    ForOfStatement as EsForOfStatement,
-    Identifier as EsIdentifier,
-    MemberExpression as EsMemberExpression,
-} from 'estree';
+import type { Expression as EsExpression, ForOfStatement as EsForOfStatement } from 'estree';
 import type { Transformer } from '../types';
-
-function getRootMemberExpression(node: EsMemberExpression): EsMemberExpression {
-    return node.object.type === 'MemberExpression' ? getRootMemberExpression(node.object) : node;
-}
-
-function getRootIdentifier(node: EsMemberExpression): EsIdentifier | null {
-    const rootMemberExpression = getRootMemberExpression(node);
-    return is.identifier(rootMemberExpression?.object) ? rootMemberExpression.object : null;
-}
 
 const bForOfYieldFrom = esTemplate`
     for (let [${is.identifier}, ${is.identifier}] of Object.entries(${is.expression} ?? {})) {
@@ -45,13 +31,7 @@ export const ForEach: Transformer<IrForEach> = function ForEach(node, cxt): EsFo
     cxt.popLocalVars();
 
     const expression = node.expression as EsExpression;
-
-    const scopeReferencedId = is.memberExpression(expression)
-        ? getRootIdentifier(expression)
-        : null;
-    const iterable = cxt.isLocalVar(scopeReferencedId?.name)
-        ? (node.expression as EsExpression)
-        : b.memberExpression(b.identifier('instance'), node.expression as EsExpression);
+    const iterable = getScopedExpression(expression, cxt);
 
     return [
         bForOfYieldFrom(
